Developer Index

This page lists the types and functions that are internal to the CFAR package. Because they are not part of the public API, these names might change relatively frequently between versions and so should not be relied upon.

All internal names are listed in the Index, and each of these entries link to the docstrings in the Docs section.

Index

This section contains a list of internal names that link to their corresponding Documentation.

Methods

CFAR.config_dir

CFAR.data_dir

CFAR.gen_gaussians

CFAR.gen_gaussians

CFAR.get_dists

CFAR.get_gaussian_config

CFAR.load_config

CFAR.results_dir

CFAR.vec_vec_to_matrix

CFAR.work_dir

Types

Constants

CFAR.ARG_CONFIG_DICT

CFAR.ARG_CONFIG_FILE

CFAR.DRWATSON_ARGS_DOC

Docs

Documentation for all internal names are listed below.

CFAR.ARG_CONFIG_DICT — Constant

ARGCONFIGDICT

Description

Common docstring: config dictionary argument.

CFAR.ARG_CONFIG_FILE — Constant

ARGCONFIGFILE

Description

Common docstring: config filename argument.

CFAR.DRWATSON_ARGS_DOC — Constant

DRWATSONARGSDOC

Description

Common docstring, the arguments to DrWatson-style directory functions.

CFAR.config_dir — Method

config_dir(args...) -> String

Summary

Points to the configs directory.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

config_dir(args)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/drwatson.jl:53.

CFAR.data_dir — Method

data_dir(args...) -> String

Summary

Points to the data directory.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

data_dir(args)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/drwatson.jl:42.

CFAR.gen_gaussians — Function

gen_gaussians(config::AbstractDict) -> Tuple{Any, Any}
gen_gaussians(
    config::AbstractDict,
    s::Float64
) -> Tuple{Any, Any}

Summary

Arguments

config::AbstractDict: the config parameters as a dictionary.

Method List / Definition Locations

gen_gaussians(config)
gen_gaussians(config, s)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/file.jl:107.

CFAR.gen_gaussians — Method

gen_gaussians(
    config_file::AbstractString
) -> Tuple{Any, Any}

Summary

Generate the Gaussian dataset from the parameters specified in the provided file.

Arguments

config_file::AbstractString: the config file name as a string.

Method List / Definition Locations

gen_gaussians(config_file)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/file.jl:157.

CFAR.get_dists — Function

get_dists(
    config::AbstractDict
) -> Vector{Distributions.MvNormal}
get_dists(
    config::AbstractDict,
    s::Float64
) -> Vector{Distributions.MvNormal}

Summary

Gets the distribution generators based upon the config parameters.

Arguments

config::AbstractDict: the config parameters as a dictionary.

Method List / Definition Locations

get_dists(config)
get_dists(config, s)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/file.jl:72.

CFAR.get_gaussian_config — Method

get_gaussian_config(config_file::AbstractString) -> Any

Summary

Loads the Gaussian distribution parameters from the provided config file.

Arguments

config_file::AbstractString: the config file name as a string.

Method List / Definition Locations

get_gaussian_config(config_file)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/file.jl:52.

CFAR.load_config — Method

load_config(config_file::AbstractString) -> Any

Summary

Wrapper for loading the configuration file with the provided filename.

Arguments

config_file::AbstractString: the config file name as a string.

Method List / Definition Locations

load_config(config_file)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/file.jl:41.

CFAR.results_dir — Method

results_dir(args...) -> String

Summary

Points to the results directory.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

results_dir(args)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/drwatson.jl:31.

CFAR.vec_vec_to_matrix — Method

vec_vec_to_matrix(vec_vec::AbstractVector) -> Any

Summary

Turns a vector of vectors into a matrix.

Assumes that the shape of the vector of vectors is square.

Arguments

vec_vec::AbstractVector: a vector of vectors of numerical values.

Method List / Definition Locations

vec_vec_to_matrix(vec_vec)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/file.jl:9.

CFAR.work_dir — Method

work_dir(args...) -> String

Summary

Points to the work directory containing raw datasets, processed datasets, and results.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

work_dir(args)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/drwatson.jl:20.