Developer Index

This page lists the types and functions that are internal to the CFAR package. Because they are not part of the public API, these names might change relatively frequently between versions and so should not be relied upon.

All internal names are listed in the Index, and each of these entries link to the docstrings in the Docs section.

Index

This section contains a list of internal names that link to their corresponding Documentation.

Methods

CFAR.config_dir

CFAR.data_dir

CFAR.gen_gaussians

CFAR.gen_gaussians

CFAR.get_dists

CFAR.get_gaussian_config

CFAR.get_mover_direction

CFAR.get_mover_line

CFAR.load_config

CFAR.results_dir

CFAR.shift_samples

CFAR.vec_vec_to_matrix

CFAR.work_dir

Types

CFAR.Data

CFAR.DataSplit

CFAR.DataSplitCombined

CFAR.DataSplitCombined

CFAR.DataSplitIndexed

CFAR.Features

CFAR.LabeledDataset

CFAR.LabeledDataset

CFAR.Labels

CFAR.MatrixData

CFAR.MoverSplit

CFAR.Targets

CFAR.VectorLabeledDataset

CFAR.VectoredData

Constants

CFAR.ARG_CONFIG_DICT

CFAR.ARG_CONFIG_FILE

CFAR.DRWATSON_ARGS_DOC

Docs

Documentation for all internal names are listed below.

CFAR.ARG_CONFIG_DICT — Constant

ARGCONFIGDICT

Description

Common docstring: config dictionary argument.

CFAR.ARG_CONFIG_FILE — Constant

ARGCONFIGFILE

Description

Common docstring: config filename argument.

CFAR.DRWATSON_ARGS_DOC — Constant

DRWATSONARGSDOC

Description

Common docstring, the arguments to DrWatson-style directory functions.

CFAR.Data — Type

Abstract supertype for all Data structs in this library.

CFAR.DataSplit — Type

A basic struct for encapsulating the components of supervised training.

CFAR.DataSplitCombined — Type

A struct for combining training and validation data, containing only train and test splits.

CFAR.DataSplitCombined — Method

Constructs a DataSplitCombined from an existing DataSplit by consolidating the training and validation data.

Arguments

data::DataSplit: the DataSplit struct for consolidating validation Features and Labels into the training data.

CFAR.DataSplitIndexed — Type

A struct for encapsulating the components of supervised training in vectorized form.

CFAR.Features — Type

Definition of features as a matrix of floating-point numbers of dimension (featuredim, nsamples).

CFAR.LabeledDataset — Type

A single dataset of Features, Targets, and human-readable string Labels.

CFAR.LabeledDataset — Method

A constructor for a LabeledDataset that merges two other LabeledDatasets.

Arguments

d1::LabeledDataset: the first LabeledDataset to consolidate.

d2::LabeledDataset: the second LabeledDataset to consolidate.

CFAR.Labels — Type

Definition of labels as a vector of strings.

CFAR.MatrixData — Type

Abstract type for Data structs that represent features as matrices.

CFAR.MoverSplit — Type

CFAR.Targets — Type

Definition of targets as a vector of integers.

CFAR.VectorLabeledDataset — Type

A single dataset of vectored labeled data with Features, Targets, and human-readable string Labels.

CFAR.VectoredData — Type

Abstract type for Data structs that represent features as vectors of matrices.

CFAR.config_dir — Method

config_dir(args...) -> String

Summary

Points to the configs directory.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

config_dir(args)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/drwatson.jl:53.

CFAR.data_dir — Method

data_dir(args...) -> String

Summary

Points to the data directory.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

data_dir(args)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/drwatson.jl:42.

CFAR.gen_gaussians — Method

gen_gaussians(config::AbstractDict) -> NTuple{4, Any}

Summary

Generates Gaussian distributed samples from the provided configuration dictionary.

Arguments

config::AbstractDict: the config parameters as a dictionary.

Method List / Definition Locations

gen_gaussians(config)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/file.jl:109.

CFAR.gen_gaussians — Method

gen_gaussians(config_file::AbstractString) -> NTuple{4, Any}

Summary

Generate the Gaussian dataset from the parameters specified in the provided file.

Arguments

config_file::AbstractString: the config file name as a string.

Method List / Definition Locations

gen_gaussians(config_file)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/file.jl:164.

CFAR.get_dists — Method

get_dists(
    config::AbstractDict
) -> Vector{Distributions.MvNormal}

Summary

Gets the distribution generators based upon the config parameters.

Arguments

config::AbstractDict: the config parameters as a dictionary.

Method List / Definition Locations

get_dists(config)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/file.jl:72.

CFAR.get_gaussian_config — Method

get_gaussian_config(config_file::AbstractString) -> Any

Summary

Loads the Gaussian distribution parameters from the provided config file.

Arguments

config_file::AbstractString: the config file name as a string.

Method List / Definition Locations

get_gaussian_config(config_file)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/file.jl:52.

CFAR.get_mover_direction — Method

get_mover_direction(config::AbstractDict) -> Any

Summary

Generates a vector representing the direction of the mover's line.

Arguments

config::AbstractDict: the config parameters as a dictionary.

Method List / Definition Locations

get_mover_direction(config)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/file.jl:177.

CFAR.get_mover_line — Method

get_mover_line(
    config::AbstractDict;
    n_points,
    length
) -> Any

Summary

Generates a dataset of points representing the mover's direction of traversal.

Arguments

config::AbstractDict: the config parameters as a dictionary.

n_points::Integer=2: kwarg, number of points along the line to return.

length::Float=10.0: kwarg, length of the line.

Method List / Definition Locations

get_mover_line(config; n_points, length)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/file.jl:215.

CFAR.load_config — Method

load_config(config_file::AbstractString) -> Any

Summary

Wrapper for loading the configuration file with the provided filename.

Arguments

config_file::AbstractString: the config file name as a string.

Method List / Definition Locations

load_config(config_file)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/file.jl:41.

CFAR.results_dir — Method

results_dir(args...) -> String

Summary

Points to the results directory.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

results_dir(args)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/drwatson.jl:31.

CFAR.shift_samples — Method

shift_samples(
    data::AbstractMatrix{T} where T<:Real,
    config::AbstractDict,
    s::Float64
) -> Any

Summary

Shifts the provided data samples matrix along the config direction by s amount.

Arguments

data::RealMatrix: the dataset of shape (n_samples, dim).

config::AbstractDict: the config parameters as a dictionary.

s::Float: the distance to travel along the line

Method List / Definition Locations

shift_samples(data, config, s)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/file.jl:196.

CFAR.vec_vec_to_matrix — Method

vec_vec_to_matrix(vec_vec::AbstractVector) -> Any

Summary

Turns a vector of vectors into a matrix.

Assumes that the shape of the vector of vectors is square.

Arguments

vec_vec::AbstractVector: a vector of vectors of numerical values.

Method List / Definition Locations

vec_vec_to_matrix(vec_vec)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/file.jl:9.

CFAR.work_dir — Method

work_dir(args...) -> String

Summary

Points to the work directory containing raw datasets, processed datasets, and results.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

work_dir(args)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/drwatson.jl:20.