Developer Index

This page lists the types and functions that are internal to the CFAR package. Because they are not part of the public API, these names might change relatively frequently between versions and so should not be relied upon.

All internal names are listed in the Index, and each of these entries link to the docstrings in the Docs section.

Index

This section contains a list of internal names that link to their corresponding Documentation.

Methods

CFAR._save_plot

CFAR._save_table

CFAR.conda_gc_disable

CFAR.conda_gc_enable

CFAR.conda_run

CFAR.conda_setup

CFAR.config_dir

CFAR.data_dir

CFAR.deserialize_df!

CFAR.deserialize_features

CFAR.dist_exp_parse

CFAR.exp_parse

CFAR.gen_gaussians

CFAR.gen_gaussians

CFAR.get_argparsesettings

CFAR.get_dists

CFAR.get_gaussian_config

CFAR.get_manual_split

CFAR.get_mlp

CFAR.get_mover_direction

CFAR.get_mover_line

CFAR.get_pylib

CFAR.get_shift

CFAR.get_table_row

CFAR.get_windows

CFAR.get_x

CFAR.get_y

CFAR.json_load

CFAR.json_save

CFAR.load_all

CFAR.load_art_sim_opts

CFAR.load_config

CFAR.load_mlp_sim_opts

CFAR.load_moversplit

CFAR.paper_results_dir

CFAR.plot_2d_attrs

CFAR.plot_2d_errlines

CFAR.plot_covellipses

CFAR.plot_mover

CFAR.results_dir

CFAR.sanitize_block_type

CFAR.sanitize_in_list

CFAR.sanitize_log_state

CFAR.save_all

CFAR.save_plot

CFAR.save_sim

CFAR.scatter_gaussian!

CFAR.setup_local_pylib

CFAR.shift_mover

CFAR.sliding_avg

CFAR.split_data

CFAR.train_test_mlp_mc

CFAR.train_test_sfam_mc

CFAR.vec_vec_to_matrix

CFAR.work_dir

Types

CFAR.ConfigDict

CFAR.Data

CFAR.DataSplit

CFAR.DataSplitCombined

CFAR.DataSplitCombined

CFAR.DataSplitIndexed

CFAR.Features

CFAR.LabeledDataset

CFAR.Labels

CFAR.MatrixData

CFAR.MoverSplit

CFAR.MoverSplit

CFAR.SerializedFeatures

CFAR.SerializedFeatures

CFAR.Targets

CFAR.VectorLabeledDataset

CFAR.VectoredData

Constants

CFAR.ARG_CONFIG_DICT

CFAR.ARG_CONFIG_FILE

CFAR.ARG_FILENAME

CFAR.ARG_P

CFAR.ARG_PLOT

CFAR.ARG_SIM_D

CFAR.ARG_SIM_DIR_FUNC

CFAR.ARG_SIM_MS

CFAR.ARG_SIM_OPTS

CFAR.BLOCK_TYPES

CFAR.COLORSCHEME

CFAR.COMMON_DOC

CFAR.DEFAULT_ART_OPTS_FILE

CFAR.DEFAULT_MLP_OPTS_FILE

CFAR.DEFAULT_N_PROCS_UNIX

CFAR.DEFAULT_N_PROCS_WINDOWS

CFAR.DPI

CFAR.DRWATSON_ARGS_DOC

CFAR.FONTFAMILY

CFAR.GRADIENTSCHEME

CFAR.JSON_INDENT

CFAR.LOCAL_PYTHON_LIBS

CFAR.LOCAL_PYTHON_LIB_LOCATION

CFAR.LOG_STATES

CFAR.MLP

CFAR.MS_GROUP

CFAR.PUBU_9

CFAR.PUBU_9_RAW

CFAR.SAVE_MAP

CFAR.SERIALIZED_FIELDS

CFAR.YLGN_9

CFAR.YLGN_9_RAW

Docs

Documentation for all internal names are listed below.

CFAR.ARG_CONFIG_DICT — Constant

ARGCONFIGDICT

Description

Common docstring: config dictionary argument.

CFAR.ARG_CONFIG_FILE — Constant

ARGCONFIGFILE

Description

Common docstring: config filename argument.

CFAR.ARG_FILENAME — Constant

ARG_FILENAME

Description

Common docstring: argument for a file name.

CFAR.ARG_P — Constant

ARG_P

Description

Common docstring: argument for a split ratio p.

CFAR.ARG_PLOT — Constant

ARG_PLOT

Description

Common docstring: argument for an existing Plots.Plot object to plot atop.

CFAR.ARG_SIM_D — Constant

ARGSIMD

Description

Common docstring: argument for the simulation options dictionary.

CFAR.ARG_SIM_DIR_FUNC — Constant

ARGSIMDIR_FUNC

Description

Common docstring: argument for a directory function

CFAR.ARG_SIM_MS — Constant

ARGSIMMS

Description

Common docstring: argument for the simulation MoverSplit dataset.

CFAR.ARG_SIM_OPTS — Constant

ARGSIMOPTS

Description

Common docstring: argument for additional simulation options.

CFAR.BLOCK_TYPES — Constant

BLOCK_TYPES

Description

The names of the blocks that are encountered during L2 experiments.

CFAR.COLORSCHEME — Constant

COLORSCHEME

Description

Plotting colorscheme.

CFAR.COMMON_DOC — Constant

COMMON_DOC

Description

Docstring prefix denoting that the constant is used as a common docstring element for other docstrings.

CFAR.DEFAULT_ART_OPTS_FILE — Constant

Location of the ART options file for simulations.

CFAR.DEFAULT_MLP_OPTS_FILE — Constant

Location of the MLP options file for simulations.

CFAR.DEFAULT_N_PROCS_UNIX — Constant

The default number of processes to start in distributed experiments on Linux.

CFAR.DEFAULT_N_PROCS_WINDOWS — Constant

The default number of processes to start in distributed experiments on Windows.

CFAR.DPI — Constant

The default plotting dots-per-inch for saving.

CFAR.DRWATSON_ARGS_DOC — Constant

DRWATSONARGSDOC

Description

Common docstring: the arguments to DrWatson-style directory functions.

CFAR.FONTFAMILY — Constant

FONTFAMILY

Description

Plotting fontfamily for all text.

CFAR.GRADIENTSCHEME — Constant

GRADIENTSCHEME

Description

Gradient scheme from a given color scheme

CFAR.JSON_INDENT — Constant

JSON_INDENT

Description

Constant for pretty indentation spacing in JSON files.

CFAR.LOCAL_PYTHON_LIBS — Constant

List of names of local libraries to make sure to install during setup.

CFAR.LOCAL_PYTHON_LIB_LOCATION — Constant

Location of the local Python libraries that are included in the project.

CFAR.LOG_STATES — Constant

LOG_STATES

Description

The enumerated states that an L2 logger log can be in.

CFAR.MLP — Constant

Name of the MLP library.

CFAR.MS_GROUP — Constant

MS_GROUP

Description

Constant name for the JLD2/H5 group that data is saved to and loaded from.

Arguments

ms::MoverSplit: the MoverSplit dataset to save.

filename::AbstractString: the full file path as a string.

CFAR.PUBU_9 — Constant

PUBU_9

Description

Purple-blue-9 color scheme.

CFAR.PUBU_9_RAW — Constant

PUBU9RAW

Description

Purple-blue-9 raw RGB values, range [0, 1].

CFAR.SAVE_MAP — Constant

SAVE_MAP

Description

Dictionary mapping the names of result save types to the private wrapper functions that implement them.

CFAR.SERIALIZED_FIELDS — Constant

SERIALIZED_FIELDS

Description

Constant declaring which fields/columns are serialized.

CFAR.YLGN_9 — Constant

YLGN_9

Description

Yellow-green-9 color scheme.

CFAR.YLGN_9_RAW — Constant

YLGN9RAW

Description

Yellow-green-9 raw RGB values, range [0, 1].

CFAR.ConfigDict — Type

ConfigDict

Description

Definition of a configuration dictionary loaded from a config file.

CFAR.Data — Type

abstract type Data

Summary

Abstract supertype for all Data structs in this library.

Fields

CFAR.DataSplit — Type

struct DataSplit <: CFAR.MatrixData

Summary

A basic struct for encapsulating the components of supervised training.

Fields

train::CFAR.LabeledDataset: Training LabeledDataset.

val::CFAR.LabeledDataset: Validation LabeledDataset.

test::CFAR.LabeledDataset: Test LabeledDataset.

CFAR.DataSplitCombined — Type

struct DataSplitCombined <: CFAR.MatrixData

Summary

A struct for combining training and validation data, containing only train and test splits.

Fields

train::CFAR.LabeledDataset: Training LabeledDataset.

test::CFAR.LabeledDataset: Testing LabeledDataset.

CFAR.DataSplitCombined — Method

DataSplitCombined(
    data::CFAR.LabeledDataset;
    p
) -> CFAR.DataSplitCombined

Summary

Returns a DataSplitCombined from a LabeledDataset with a provided split ratio p.

Arguments

data::LabeledDataset: the original LabeledDataset to split into a DataSplitCombined.

p::Float=0.8: kwarg, the split ratio ∈ (0, 1).

Method List / Definition Locations

DataSplitCombined(data; p)

defined at /home/runner/work/CFAR/CFAR/src/lib/data.jl:201.

CFAR.DataSplitIndexed — Type

struct DataSplitIndexed <: CFAR.VectoredData

Summary

A struct for encapsulating the components of supervised training in vectorized form.

Fields

train::CFAR.VectorLabeledDataset: Training VectorLabeledDataset.

val::CFAR.VectorLabeledDataset: Validation VectorLabeledDataset.

test::CFAR.VectorLabeledDataset: Test VectorLabeledDataset.

CFAR.Features — Type

Features

Description

Definition of features as a matrix of floating-point numbers of dimension (featuredim, nsamples).

CFAR.LabeledDataset — Type

struct LabeledDataset

Summary

A single dataset of Features, Targets, and human-readable string Labels.

Fields

x::ElasticArrays.ElasticMatrix{Float64, V} where V<:DenseVector{Float64}: Collection of Features in the labeled dataset.

y::ElasticArrays.ElasticVector{Int64, V} where V<:DenseVector{Int64}: Targets corresponding to the Features.

labels::ElasticArrays.ElasticVector{String, V} where V<:DenseVector{String}: Human-readable Labels corresponding to the Targets values.

CFAR.Labels — Type

Labels

Description

Definition of labels as a vector of strings.

CFAR.MatrixData — Type

abstract type MatrixData <: CFAR.Data

Summary

Abstract type for Data structs that represent features as matrices.

Fields

CFAR.MoverSplit — Type

struct MoverSplit

Summary

Definition of a split of data with one part remaining static and the other moving.

Fields

static::CFAR.DataSplitCombined: The static DataSplitCombined.

mover::CFAR.DataSplitCombined: The moving DataSplitCombined.

config::Dict{Any, Any}: The config used to generate the mover.

CFAR.MoverSplit — Method

MoverSplit(
    static::CFAR.LabeledDataset,
    mover::CFAR.LabeledDataset,
    config::Dict{Any, Any}
) -> CFAR.MoverSplit

Summary

Constructor for a MoverSplit taking a preconstructed static and mover LabeledDatasets along with a split parameter p.

Arguments

static::LabeledDataset: the static part of the dataset.

mover::LabeledDataset: the moving part of the datset.

Method List / Definition Locations

MoverSplit(static, mover, config)

defined at /home/runner/work/CFAR/CFAR/src/lib/data.jl:238.

CFAR.SerializedFeatures — Type

struct SerializedFeatures

Summary

Serializer for Features for saving with Arrow.

Fields

dim1::Vector{Float64}: The first dimension of the Features.

dim2::Vector{Float64}: The second dimension of the Features.

CFAR.SerializedFeatures — Method

SerializedFeatures(
    data::ElasticArrays.ElasticMatrix{Float64, V} where V<:DenseVector{Float64}
) -> CFAR.SerializedFeatures

Summary

Constructs a SerializedFeatures from a set of Features.

Arguments

data::Features: the Features to serialize.

Method List / Definition Locations

SerializedFeatures(data)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:333.

CFAR.Targets — Type

Targets

Description

Definition of targets as a vector of integers.

CFAR.VectorLabeledDataset — Type

struct VectorLabeledDataset

Summary

A single dataset of vectored labeled data with Features, Targets, and human-readable string Labels.

Fields

x::Vector{ElasticArrays.ElasticMatrix{Float64, V} where V<:DenseVector{Float64}}: A vector of Features matrices.

y::Vector{ElasticArrays.ElasticVector{Int64, V} where V<:DenseVector{Int64}}: A vector of Targets corresponding to the Features.

labels::ElasticArrays.ElasticVector{String, V} where V<:DenseVector{String}: String Labels corresponding to the Targets.

CFAR.VectoredData — Type

abstract type VectoredData <: CFAR.Data

Summary

Abstract type for Data structs that represent features as vectors of matrices.

Fields

CFAR._save_plot — Method

_save_plot(p::Plots.Plot, filename::AbstractString) -> Any

Summary

Wrapper for how figures are saved in the CFAR project.

Arguments

p::Plots.Plot: the Plot object to save.

filename::AbstractString: the full file path as a string.

Method List / Definition Locations

_save_plot(p, filename)

defined at /home/runner/work/CFAR/CFAR/src/lib/plot.jl:330.

CFAR._save_table — Method

_save_table(table, filename::AbstractString) -> Any

Summary

Wrapper for how tables are saved in the CFAR project.

Arguments

table: the table object to save.

filename::AbstractString: the full file path as a string.

Method List / Definition Locations

_save_table(table, filename)

defined at /home/runner/work/CFAR/CFAR/src/lib/plot.jl:341.

CFAR.conda_gc_disable — Method

conda_gc_disable()

Summary

Wrapper for disabling the PythonCall garbage collector.

Method List / Definition Locations

conda_gc_disable()

defined at /home/runner/work/CFAR/CFAR/src/lib/py.jl:98.

CFAR.conda_gc_enable — Method

conda_gc_enable()

Summary

Wrapper for reenabling the PythonCall garbage collector.

Method List / Definition Locations

conda_gc_enable()

defined at /home/runner/work/CFAR/CFAR/src/lib/py.jl:109.

CFAR.conda_run — Method

conda_run(cmd_string::AbstractString) -> Any

Summary

Runs a provided command with the correct CondaPkg.jl Python environment.

Arguments

cmd_string::AbstractString: the Python command to run as a string, excluding the initial 'python' part.

Method List / Definition Locations

conda_run(cmd_string)

defined at /home/runner/work/CFAR/CFAR/src/lib/py.jl:43.

CFAR.conda_setup — Method

conda_setup()

Summary

Sets up the Conda dependencies, including local libraries.

Method List / Definition Locations

conda_setup()

defined at /home/runner/work/CFAR/CFAR/src/lib/py.jl:81.

CFAR.config_dir — Method

config_dir(args...) -> String

Summary

Points to the configs directory.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

config_dir(args)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/drwatson.jl:56.

CFAR.data_dir — Method

data_dir(args...) -> String

Summary

Points to the data directory.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

data_dir(args)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/drwatson.jl:44.

CFAR.deserialize_df! — Method

deserialize_df!(df::DataFrames.DataFrame)

Summary

Deserializes the serialized fields of a DataFrame according to SERIALIZED_FIELDS.

Arguments

df::DataFrames.DataFrame: the dataframe containing serialized fields.

Method List / Definition Locations

deserialize_df!(df)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:433.

CFAR.deserialize_features — Method

deserialize_features(
    el::CFAR.SerializedFeatures
) -> ElasticArrays.ElasticMatrix{Float64, Vector{Float64}}

Summary

Deserializes a set of SerializedFeatures and constructs a set of Features.

Arguments

el::SerializedFeatures: the SerializedFeatures to deserialize.

Method List / Definition Locations

deserialize_features(el)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:407.

CFAR.dist_exp_parse — Function

dist_exp_parse() -> Any
dist_exp_parse(description::AbstractString) -> Any

Summary

Parses the command line for common options in distributed experiments.

Arguments

description::AbstractString: optional positional, the script description for the parser

Method List / Definition Locations

dist_exp_parse()
dist_exp_parse(description)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/args.jl:61.

CFAR.exp_parse — Function

exp_parse() -> Any
exp_parse(description::AbstractString) -> Any

Summary

Parses the command line for common options in serial (non-distributed) experiments.

Arguments

description::AbstractString: optional positional, the script description for the parser

Method List / Definition Locations

exp_parse()
exp_parse(description)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/args.jl:35.

CFAR.gen_gaussians — Method

gen_gaussians(
    config_file::AbstractString
) -> CFAR.MoverSplit

Summary

Generate the Gaussian dataset from the parameters specified in the provided file.

Arguments

config_file::AbstractString: the config file name as a string.

Method List / Definition Locations

gen_gaussians(config_file)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:174.

CFAR.gen_gaussians — Method

gen_gaussians(config::Dict{Any, Any}) -> CFAR.MoverSplit

Summary

Generates Gaussian distributed samples from the provided configuration dictionary.

Arguments

config::ConfigDict: the config parameters as a dictionary.

Method List / Definition Locations

gen_gaussians(config)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:113.

CFAR.get_argparsesettings — Function

get_argparsesettings() -> Any
get_argparsesettings(description::AbstractString) -> Any

Summary

Common function for how ArgParseSettings are generated in the project.

Arguments

description::AbstractString: optional positional, the script description for the parser

Method List / Definition Locations

get_argparsesettings()
get_argparsesettings(description)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/args.jl:19.

CFAR.get_dists — Method

get_dists(
    config::Dict{Any, Any}
) -> Vector{Distributions.MvNormal}

Summary

Gets the distribution generators based upon the config parameters.

Arguments

config::ConfigDict: the config parameters as a dictionary.

Method List / Definition Locations

get_dists(config)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:76.

CFAR.get_gaussian_config — Method

get_gaussian_config(config_file::AbstractString) -> Any

Summary

Loads the Gaussian distribution parameters from the provided config file.

Arguments

config_file::AbstractString: the config file name as a string.

Method List / Definition Locations

get_gaussian_config(config_file)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:37.

CFAR.get_manual_split — Method

get_manual_split(
    data::AbstractMatrix{T} where T<:Real,
    targets::AbstractVector{T} where T<:Integer;
    p
) -> Tuple{Tuple{Any, Any}, Tuple{Any, Any}}

Summary

Teturns a manual train/test x/y split from a data matrix and labels using MLDataUtils.

Arguments

data::RealMatrix: the feature data to split into training and testing.

targets::IntegerVector: the labels corresponding to the data to split into training and testing.

p::Float=0.8: kwarg, the split ratio ∈ (0, 1).

Method List / Definition Locations

get_manual_split(data, targets; p)

defined at /home/runner/work/CFAR/CFAR/src/lib/data.jl:161.

CFAR.get_mlp — Method

get_mlp() -> PythonCall.Py

Summary

Loads and returns a handle to the local mlp Python library.

Method List / Definition Locations

get_mlp()

defined at /home/runner/work/CFAR/CFAR/src/lib/py.jl:32.

CFAR.get_mover_direction — Method

get_mover_direction(config::Dict{Any, Any}) -> Any

Summary

Generates a vector representing the direction of the mover's line.

Arguments

config::ConfigDict: the config parameters as a dictionary.

Method List / Definition Locations

get_mover_direction(config)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:188.

CFAR.get_mover_line — Method

get_mover_line(
    config::Dict{Any, Any};
    n_points,
    length
) -> Any

Summary

Generates a dataset of points representing the mover's direction of traversal.

Arguments

config::ConfigDict: the config parameters as a dictionary.

n_points::Integer=2: kwarg, number of points along the line to return.

length::Float=10.0: kwarg, length of the line.

Method List / Definition Locations

get_mover_line(config; n_points, length)

defined at /home/runner/work/CFAR/CFAR/src/lib/plot.jl:78.

CFAR.get_pylib — Method

get_pylib(lib::AbstractString) -> PythonCall.Py

Summary

Loads and returns a handle to the provided local Python library.

Arguments

lib::AbstractString: the string name of the local Python library to load.

Method List / Definition Locations

get_pylib(lib)

defined at /home/runner/work/CFAR/CFAR/src/lib/py.jl:17.

CFAR.get_shift — Method

get_shift(config::Dict{Any, Any}, s::Float64) -> Any

Summary

Gets the shift vector from the configuration and distance to traverse.

Arguments

config::ConfigDict: the config parameters as a dictionary.

s::Float: the distance to travel along the line

Method List / Definition Locations

get_shift(config, s)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:206.

CFAR.get_table_row — Method

get_table_row(
    data::CFAR.DataSplitCombined,
    label::AbstractString
) -> Any

Summary

Constructs a DataFrame table row for saving to an Arrow table.

Arguments

data::DataSplitCombined: the data split.

label::AbstractString: the string label for the data split.

Method List / Definition Locations

get_table_row(data, label)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:351.

CFAR.get_windows — Method

get_windows(
    vs::AbstractVector{T} where T<:Real,
    n::Integer
) -> Any

Summary

Constructs a windowed matrix of a vector.

Arguments

vs::RealVector: the original vector.

n::Integer: the size of the sliding window.

Method List / Definition Locations

get_windows(vs, n)

defined at /home/runner/work/CFAR/CFAR/src/lib/plot.jl:260.

CFAR.get_x — Method

get_x(
    data::CFAR.DataSplitCombined
) -> ElasticArrays.ElasticMatrix{Float64, _A} where _A<:DenseVector{Float64}

Summary

Returns the concatenated training and testing features data from a DataSplitCombined.

Arguments

data::DataSplitCombined: the DataSplitCombined to get the combined train/test features from.

Method List / Definition Locations

get_x(data)

defined at /home/runner/work/CFAR/CFAR/src/lib/data.jl:258.

CFAR.get_y — Method

get_y(data::CFAR.DataSplitCombined) -> AbstractVector

Summary

Returns the concatenated training and testing targets data from a DataSplitCombined.

Arguments

data::DataSplitCombined: the DataSplitCombined to get the combined train/test targets from.

Method List / Definition Locations

get_y(data)

defined at /home/runner/work/CFAR/CFAR/src/lib/data.jl:271.

CFAR.json_load — Method

json_load(filepath::AbstractString) -> Any

Summary

Loads the JSON file, interpreted as a dictionary.

Arguments

filepath::AbstractString: the full file name (with path) to load.

Method List / Definition Locations

json_load(filepath)

defined at /home/runner/work/CFAR/CFAR/src/lib/l2/common.jl:96.

CFAR.json_save — Method

json_save(
    filepath::AbstractString,
    dict::AbstractDict
) -> Any

Summary

Saves the dictionary to a JSON file.

Arguments

filepath::AbstractString: the full file name (with path) to save to.

dict::AbstractDict: the dictionary to save to the file.

Method List / Definition Locations

json_save(filepath, dict)

defined at /home/runner/work/CFAR/CFAR/src/lib/l2/common.jl:83.

CFAR.load_all — Method

load_all(
    filename::AbstractString
) -> Tuple{DataFrames.DataFrame, Arrow.Table}

Summary

Loads the Arrow file as a DataFrame.

Arguments

filename::AbstractString: location of the Arrow file.

Method List / Definition Locations

load_all(filename)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:447.

CFAR.load_art_sim_opts — Function

load_art_sim_opts() -> Any
load_art_sim_opts(file::AbstractString) -> Any

Summary

Loads the ART simulation options in the provided file with a default.

Arguments

file::AbstractString: the YAML file to load, default art.yml.

Method List / Definition Locations

load_art_sim_opts()
load_art_sim_opts(file)

defined at /home/runner/work/CFAR/CFAR/src/lib/experiments.jl:17.

CFAR.load_config — Method

load_config(config_file::AbstractString) -> Any

Summary

Wrapper for loading the configuration file with the provided filename.

Arguments

config_file::AbstractString: the config file name as a string.

Method List / Definition Locations

load_config(config_file)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/file.jl:28.

CFAR.load_mlp_sim_opts — Function

load_mlp_sim_opts() -> Any
load_mlp_sim_opts(file::AbstractString) -> Any

Summary

Loads the MLP simulation options in the provided file with a default.

Arguments

file::AbstractString: the YAML file to load, default mlp.yml.

Method List / Definition Locations

load_mlp_sim_opts()
load_mlp_sim_opts(file)

defined at /home/runner/work/CFAR/CFAR/src/lib/experiments.jl:42.

CFAR.load_moversplit — Method

load_moversplit(filename::AbstractString) -> Any

Summary

Loads and returns the gaussian data from the provided filename.

Arguments

filename::AbstractString: the full file path as a string.

Method List / Definition Locations

load_moversplit(filename)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:307.

CFAR.paper_results_dir — Method

paper_results_dir(args...) -> String

Summary

DrWatson-style paper results directory.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

paper_results_dir(args)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/drwatson.jl:68.

CFAR.plot_2d_attrs — Method

plot_2d_attrs(
    df::DataFrames.DataFrame,
    attrs::Array{T<:AbstractString, 1};
    avg,
    n,
    title
) -> Plots.Plot

Summary

Plots the 2D performances trends.

Arguments

df::DataFrame: the collected simulation results.

attrs::Vector{T} where T <: AbstractString: the columns in the dataframe as a list of strings to create plotlines for.

avg::Bool=false: optional, default false, flag to compute the windowed averages of the trends.

n::Integer=10: optional, default 10, the size of the average sliding window if that option is used.

Method List / Definition Locations

plot_2d_attrs(df, attrs; avg, n, title)

defined at /home/runner/work/CFAR/CFAR/src/lib/plot.jl:206.

CFAR.plot_2d_errlines — Method

plot_2d_errlines(
    df::DataFrames.DataFrame,
    attrs::Array{T<:AbstractString, 1};
    n,
    title
) -> Plots.Plot

Summary

Plots the 2D performances trends.

Arguments

df::DataFrame: the collected simulation results.

attrs::Vector{T} where T <: AbstractString: the columns in the dataframe as a list of strings to create plotlines for.

Method List / Definition Locations

plot_2d_errlines(df, attrs; n, title)

defined at /home/runner/work/CFAR/CFAR/src/lib/plot.jl:283.

CFAR.plot_covellipses — Method

plot_covellipses(p::Plots.Plot, config::Dict{Any, Any})

Summary

Plots the covariance ellipses from the config on top of an existing plot.

Arguments

p::Plots.Plot: an existing Plots.Plot object.

config::ConfigDict: the config parameters as a dictionary.

Method List / Definition Locations

plot_covellipses(p, config)

defined at /home/runner/work/CFAR/CFAR/src/lib/plot.jl:47.

CFAR.plot_mover — Method

plot_mover(ms::CFAR.MoverSplit) -> Plots.Plot

Summary

Plots the mover line plot with scattered data points, covariance lines, and mover line.

Arguments

ms::MoverSplit: the MoverSplit dataset.

Method List / Definition Locations

plot_mover(ms)

defined at /home/runner/work/CFAR/CFAR/src/lib/plot.jl:105.

CFAR.results_dir — Method

results_dir(args...) -> String

Summary

Points to the results directory.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

results_dir(args)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/drwatson.jl:32.

CFAR.sanitize_block_type — Method

sanitize_block_type(block_type::AbstractString)

Summary

Sanitize the selected block type against the BLOCK_TYPES constant.

Arguments

block_type::AbstractString: the selected block type.

Method List / Definition Locations

sanitize_block_type(block_type)

defined at /home/runner/work/CFAR/CFAR/src/lib/l2/common.jl:60.

CFAR.sanitize_in_list — Method

sanitize_in_list(
    selection_type::AbstractString,
    selection,
    acceptable::Array{T, 1}
)

Summary

Sanitizes a selection within a list of acceptable options.

Arguments

selection_type::AbstractString: a string describing the option in case it is misused.

selection::Any: a single selection from a list.

Method List / Definition Locations

sanitize_in_list(selection_type, selection, acceptable)

defined at /home/runner/work/CFAR/CFAR/src/lib/l2/common.jl:45.

CFAR.sanitize_log_state — Method

sanitize_log_state(log_state::AbstractString)

Summary

Sanitize the selected log state against the LOG_STATES constant.

Arguments

log_state::AbstractString: the selected log state.

Method List / Definition Locations

sanitize_log_state(log_state)

defined at /home/runner/work/CFAR/CFAR/src/lib/l2/common.jl:71.

CFAR.save_all — Method

save_all(
    ms::CFAR.MoverSplit,
    filename::AbstractString
) -> DataFrames.DataFrame

Summary

Saves the MoverSplit as an Arrow file for transferability.

Arguments

ms::MoverSplit:

Method List / Definition Locations

save_all(ms, filename)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:368.

CFAR.save_plot — Method

save_plot(
    p::Plots.Plot,
    fig_name::AbstractString,
    exp_top::AbstractString,
    exp_name::AbstractString
)

Summary

Saves the plot to the both the local results directory and to the paper directory.

Arguments

p::Plots.Plot: the handle of the plot to save.

fig_name::AbstractString: the name of the figure file itself.

exp_top::AbstractString: the top of the experiment directory.

exp_name::AbstractString: the name of the experiment itself.

Method List / Definition Locations

save_plot(p, fig_name, exp_top, exp_name)

defined at /home/runner/work/CFAR/CFAR/src/lib/plot.jl:364.

CFAR.save_sim — Method

save_sim(
    dir_func::Function,
    d::AbstractDict,
    fulld::AbstractDict
)

Summary

Common save function for simulations.

Arguments

dir_func::Function: the function that provides the correct file path with provided strings.

d::AbstractDict: the simulation options dictionary.

fulld::AbstractDict: the dictionary containing the sim results.

Method List / Definition Locations

save_sim(dir_func, d, fulld)

defined at /home/runner/work/CFAR/CFAR/src/lib/experiments.jl:64.

CFAR.scatter_gaussian! — Method

scatter_gaussian!(
    p::Plots.Plot,
    data::CFAR.DataSplitCombined
)

Summary

Combines and plots data from a gaussian distribution.

Arguments

p::Plots.Plot: the plot handle to add the gaussians to.

data::DataSplitCombined: the data to plot.

Method List / Definition Locations

scatter_gaussian!(p, data)

defined at /home/runner/work/CFAR/CFAR/src/lib/plot.jl:18.

CFAR.setup_local_pylib — Method

setup_local_pylib(lib::AbstractString)

Summary

Sets up a local Python library with .

Arguments

lib::AbstractString: the string name of the local library to setup.

Method List / Definition Locations

setup_local_pylib(lib)

defined at /home/runner/work/CFAR/CFAR/src/lib/py.jl:61.

CFAR.shift_mover — Method

shift_mover(
    ms::CFAR.MoverSplit,
    s::Float64
) -> CFAR.MoverSplit

Summary

Moves the mover component of a MoverSplit a distance of s.

Arguments

ms::MoverSplit: the datset containing a mover to shift.

config::ConfigDict: the config parameters as a dictionary.

s::Float: the distance to travel along the line

Method List / Definition Locations

shift_mover(ms, s)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:241.

CFAR.sliding_avg — Method

sliding_avg(
    vs::AbstractVector{T} where T<:Real,
    n::Integer
) -> Any

Summary

Computes the averages of a sliding window along a vector.

Arguments

vs::RealVector: the vector to compute windowed averages of.

n::Integer: the size of the sliding window.

Method List / Definition Locations

sliding_avg(vs, n)

defined at /home/runner/work/CFAR/CFAR/src/lib/plot.jl:187.

CFAR.split_data — Method

split_data(
    data::CFAR.LabeledDataset;
    p
) -> Tuple{CFAR.LabeledDataset, CFAR.LabeledDataset}

Summary

Splits a LabeledDataset into two LabeledDatasets

Arguments

data::LabeledDataset: the original dataset to split.

p::Float=0.8: kwarg, the split ratio ∈ (0, 1).

Method List / Definition Locations

split_data(data; p)

defined at /home/runner/work/CFAR/CFAR/src/lib/data.jl:175.

CFAR.train_test_mlp_mc — Method

train_test_mlp_mc(
    d::AbstractDict,
    ms::CFAR.MoverSplit,
    dir_func::Function,
    opts::AbstractDict
)

Summary

Train and test an MLP on the MoverSplit dataset.

Arguments

d::AbstractDict: the simulation options dictionary.

ms::MoverSplit: the MoverSplit dataset to train and test on.

dir_func::Function: the function that provides the correct file path with provided strings.

opts::AbstractDict: additional options for the simulation.

Method List / Definition Locations

train_test_mlp_mc(d, ms, dir_func, opts)

defined at /home/runner/work/CFAR/CFAR/src/lib/experiments.jl:202.

CFAR.train_test_sfam_mc — Method

train_test_sfam_mc(
    d::AbstractDict,
    ms::CFAR.MoverSplit,
    dir_func::Function,
    opts::AbstractDict
)

Summary

Train and test SFAM on the MoverSplit dataset in parallel.

Arguments

d::AbstractDict: the simulation options dictionary.

ms::MoverSplit: the MoverSplit dataset to train and test on.

dir_func::Function: the function that provides the correct file path with provided strings.

opts::AbstractDict: additional options for the simulation.

Method List / Definition Locations

train_test_sfam_mc(d, ms, dir_func, opts)

defined at /home/runner/work/CFAR/CFAR/src/lib/experiments.jl:91.

CFAR.vec_vec_to_matrix — Method

vec_vec_to_matrix(vec_vec::AbstractVector) -> Any

Summary

Turns a vector of vectors into a matrix.

Assumes that the shape of the vector of vectors is square.

Arguments

vec_vec::AbstractVector: a vector of vectors of numerical values.

Method List / Definition Locations

vec_vec_to_matrix(vec_vec)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:19.

CFAR.work_dir — Method

work_dir(args...) -> String

Summary

Points to the work directory containing raw datasets, processed datasets, and results.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

work_dir(args)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/drwatson.jl:20.