Developer Index

This page lists the types and functions that are internal to the CFAR package. Because they are not part of the public API, these names might change relatively frequently between versions and so should not be relied upon.

All internal names are listed in the Index, and each of these entries link to the docstrings in the Docs section.

Index

This section contains a list of internal names that link to their corresponding Documentation.

Methods

CFAR.config_dir

CFAR.data_dir

CFAR.gen_gaussians

CFAR.gen_gaussians

CFAR.get_dists

CFAR.get_gaussian_config

CFAR.get_manual_split

CFAR.get_mover_direction

CFAR.get_mover_line

CFAR.get_x

CFAR.get_y

CFAR.load_config

CFAR.results_dir

CFAR.scatter_gaussian!

CFAR.shift_samples

CFAR.split_data

CFAR.vec_vec_to_matrix

CFAR.work_dir

Types

CFAR.ConfigDict

CFAR.Data

CFAR.DataSplit

CFAR.DataSplitCombined

CFAR.DataSplitCombined

CFAR.DataSplitIndexed

CFAR.Features

CFAR.LabeledDataset

CFAR.Labels

CFAR.MatrixData

CFAR.MoverSplit

CFAR.MoverSplit

CFAR.Targets

CFAR.VectorLabeledDataset

CFAR.VectoredData

Constants

CFAR.ARG_CONFIG_DICT

CFAR.ARG_CONFIG_FILE

CFAR.ARG_P

CFAR.DRWATSON_ARGS_DOC

Docs

Documentation for all internal names are listed below.

CFAR.ARG_CONFIG_DICT — Constant

ARGCONFIGDICT

Description

Common docstring: config dictionary argument.

CFAR.ARG_CONFIG_FILE — Constant

ARGCONFIGFILE

Description

Common docstring: config filename argument.

CFAR.ARG_P — Constant

ARG_P

Description

Common docstring: argument for a split ratio p.

CFAR.DRWATSON_ARGS_DOC — Constant

DRWATSONARGSDOC

Description

Common docstring, the arguments to DrWatson-style directory functions.

CFAR.ConfigDict — Type

ConfigDict

Description

Definition of a configuration dictionary loaded from a config file.

CFAR.Data — Type

abstract type Data

Summary

Abstract supertype for all Data structs in this library.

Fields

CFAR.DataSplit — Type

struct DataSplit <: CFAR.MatrixData

Summary

A basic struct for encapsulating the components of supervised training.

Fields

train::CFAR.LabeledDataset: Training LabeledDataset.

val::CFAR.LabeledDataset: Validation LabeledDataset.

test::CFAR.LabeledDataset: Test LabeledDataset.

CFAR.DataSplitCombined — Type

struct DataSplitCombined <: CFAR.MatrixData

Summary

A struct for combining training and validation data, containing only train and test splits.

Fields

train::CFAR.LabeledDataset: Training LabeledDataset.

test::CFAR.LabeledDataset: Testing LabeledDataset.

CFAR.DataSplitCombined — Method

DataSplitCombined(
    data::CFAR.LabeledDataset;
    p
) -> CFAR.DataSplitCombined

Summary

Returns a DataSplitCombined from a LabeledDataset with a provided split ratio p.

Arguments

data::LabeledDataset: the original LabeledDataset to split into a DataSplitCombined.

p::Float=0.8: kwarg, the split ratio ∈ (0, 1).

Method List / Definition Locations

DataSplitCombined(data; p)

defined at /home/runner/work/CFAR/CFAR/src/lib/data.jl:196.

CFAR.DataSplitIndexed — Type

struct DataSplitIndexed <: CFAR.VectoredData

Summary

A struct for encapsulating the components of supervised training in vectorized form.

Fields

train::CFAR.VectorLabeledDataset: Training VectorLabeledDataset.

val::CFAR.VectorLabeledDataset: Validation VectorLabeledDataset.

test::CFAR.VectorLabeledDataset: Test VectorLabeledDataset.

CFAR.Features — Type

Features

Description

Definition of features as a matrix of floating-point numbers of dimension (featuredim, nsamples).

CFAR.LabeledDataset — Type

struct LabeledDataset

Summary

A single dataset of Features, Targets, and human-readable string Labels.

Fields

x::ElasticArrays.ElasticMatrix{Float64, V} where V<:DenseVector{Float64}: Collection of Features in the labeled dataset.

y::ElasticArrays.ElasticVector{Int64, V} where V<:DenseVector{Int64}: Targets corresponding to the Features.

labels::ElasticArrays.ElasticVector{String, V} where V<:DenseVector{String}: Human-readable Labels corresponding to the Targets values.

CFAR.Labels — Type

Labels

Description

Definition of labels as a vector of strings.

CFAR.MatrixData — Type

abstract type MatrixData <: CFAR.Data

Summary

Abstract type for Data structs that represent features as matrices.

Fields

CFAR.MoverSplit — Type

struct MoverSplit <: CFAR.MatrixData

Summary

Definition of a split of data with one part remaining static and the other moving.

Fields

static::CFAR.DataSplitCombined: The static DataSplitCombined.

mover::CFAR.DataSplitCombined: The moving DataSplitCombined.

CFAR.MoverSplit — Method

MoverSplit(
    static::CFAR.LabeledDataset,
    mover::CFAR.LabeledDataset;
    p
) -> CFAR.MoverSplit

Summary

Constructor for a MoverSplit taking a preconstructed static and mover LabeledDatasets along with a split parameter p.

Method List / Definition Locations

MoverSplit(static, mover; p)

defined at /home/runner/work/CFAR/CFAR/src/lib/data.jl:222.

CFAR.Targets — Type

Targets

Description

Definition of targets as a vector of integers.

CFAR.VectorLabeledDataset — Type

struct VectorLabeledDataset

Summary

A single dataset of vectored labeled data with Features, Targets, and human-readable string Labels.

Fields

x::Vector{ElasticArrays.ElasticMatrix{Float64, V} where V<:DenseVector{Float64}}: A vector of Features matrices.

y::Vector{ElasticArrays.ElasticVector{Int64, V} where V<:DenseVector{Int64}}: A vector of Targets corresponding to the Features.

labels::ElasticArrays.ElasticVector{String, V} where V<:DenseVector{String}: String Labels corresponding to the Targets.

CFAR.VectoredData — Type

abstract type VectoredData <: CFAR.Data

Summary

Abstract type for Data structs that represent features as vectors of matrices.

Fields

CFAR.config_dir — Method

config_dir(args...) -> String

Summary

Points to the configs directory.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

config_dir(args)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/drwatson.jl:53.

CFAR.data_dir — Method

data_dir(args...) -> String

Summary

Points to the data directory.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

data_dir(args)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/drwatson.jl:42.

CFAR.gen_gaussians — Method

gen_gaussians(
    config_file::AbstractString
) -> CFAR.MoverSplit

Summary

Generate the Gaussian dataset from the parameters specified in the provided file.

Arguments

config_file::AbstractString: the config file name as a string.

Method List / Definition Locations

gen_gaussians(config_file)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:152.

CFAR.gen_gaussians — Method

gen_gaussians(config::Dict{Any, Any}) -> CFAR.MoverSplit

Summary

Generates Gaussian distributed samples from the provided configuration dictionary.

Arguments

config::ConfigDict: the config parameters as a dictionary.

Method List / Definition Locations

gen_gaussians(config)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:91.

CFAR.get_dists — Method

get_dists(
    config::Dict{Any, Any}
) -> Vector{Distributions.MvNormal}

Summary

Gets the distribution generators based upon the config parameters.

Arguments

config::ConfigDict: the config parameters as a dictionary.

Method List / Definition Locations

get_dists(config)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:54.

CFAR.get_gaussian_config — Method

get_gaussian_config(config_file::AbstractString) -> Any

Summary

Loads the Gaussian distribution parameters from the provided config file.

Arguments

config_file::AbstractString: the config file name as a string.

Method List / Definition Locations

get_gaussian_config(config_file)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:34.

CFAR.get_manual_split — Method

get_manual_split(
    data::AbstractMatrix{T} where T<:Real,
    targets::AbstractVector{T} where T<:Integer;
    p
) -> Tuple{Tuple{Any, Any}, Tuple{Any, Any}}

Summary

Teturns a manual train/test x/y split from a data matrix and labels using MLDataUtils.

Arguments

data::RealMatrix: the feature data to split into training and testing.

targets::IntegerVector: the labels corresponding to the data to split into training and testing.

p::Float=0.8: kwarg, the split ratio ∈ (0, 1).

Method List / Definition Locations

get_manual_split(data, targets; p)

defined at /home/runner/work/CFAR/CFAR/src/lib/data.jl:158.

CFAR.get_mover_direction — Method

get_mover_direction(config::Dict{Any, Any}) -> Any

Summary

Generates a vector representing the direction of the mover's line.

Arguments

config::ConfigDict: the config parameters as a dictionary.

Method List / Definition Locations

get_mover_direction(config)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:165.

CFAR.get_mover_line — Method

get_mover_line(
    config::Dict{Any, Any};
    n_points,
    length
) -> Any

Summary

Generates a dataset of points representing the mover's direction of traversal.

Arguments

config::ConfigDict: the config parameters as a dictionary.

n_points::Integer=2: kwarg, number of points along the line to return.

length::Float=10.0: kwarg, length of the line.

Method List / Definition Locations

get_mover_line(config; n_points, length)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:203.

CFAR.get_x — Method

get_x(
    data::CFAR.DataSplitCombined
) -> ElasticArrays.ElasticMatrix{Float64, _A} where _A<:DenseVector{Float64}

Summary

Returns the concatenated training and testing features data from a DataSplitCombined.

Arguments

data::DataSplitCombined: the DataSplitCombined to get the combined train/test features from.

Method List / Definition Locations

get_x(data)

defined at /home/runner/work/CFAR/CFAR/src/lib/data.jl:237.

CFAR.get_y — Method

get_y(data::CFAR.DataSplitCombined) -> AbstractVector

Summary

Returns the concatenated training and testing targets data from a DataSplitCombined.

Arguments

data::DataSplitCombined: the DataSplitCombined to get the combined train/test targets from.

Method List / Definition Locations

get_y(data)

defined at /home/runner/work/CFAR/CFAR/src/lib/data.jl:250.

CFAR.load_config — Method

load_config(config_file::AbstractString) -> Any

Summary

Wrapper for loading the configuration file with the provided filename.

Arguments

config_file::AbstractString: the config file name as a string.

Method List / Definition Locations

load_config(config_file)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/file.jl:20.

CFAR.results_dir — Method

results_dir(args...) -> String

Summary

Points to the results directory.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

results_dir(args)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/drwatson.jl:31.

CFAR.scatter_gaussian! — Method

scatter_gaussian!(
    p::Plots.Plot,
    data::CFAR.DataSplitCombined
)

Summary

Combines and plots data from a gaussian distribution.

Method List / Definition Locations

scatter_gaussian!(p, data)

defined at /home/runner/work/CFAR/CFAR/src/lib/plot.jl:11.

CFAR.shift_samples — Method

shift_samples(
    data::AbstractMatrix{T} where T<:Real,
    config::Dict{Any, Any},
    s::Float64
) -> Any

Summary

Shifts the provided data samples matrix along the config direction by s amount.

Arguments

data::RealMatrix: the dataset of shape (n_samples, dim).

config::ConfigDict: the config parameters as a dictionary.

s::Float: the distance to travel along the line

Method List / Definition Locations

shift_samples(data, config, s)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:184.

CFAR.split_data — Method

split_data(
    data::CFAR.LabeledDataset;
    p
) -> Tuple{CFAR.LabeledDataset, CFAR.LabeledDataset}

Summary

Splits a LabeledDataset into two LabeledDatasets

Arguments

data::LabeledDataset: the original dataset to split.

p::Float=0.8: kwarg, the split ratio ∈ (0, 1).

Method List / Definition Locations

split_data(data; p)

defined at /home/runner/work/CFAR/CFAR/src/lib/data.jl:170.

CFAR.vec_vec_to_matrix — Method

vec_vec_to_matrix(vec_vec::AbstractVector) -> Any

Summary

Turns a vector of vectors into a matrix.

Assumes that the shape of the vector of vectors is square.

Arguments

vec_vec::AbstractVector: a vector of vectors of numerical values.

Method List / Definition Locations

vec_vec_to_matrix(vec_vec)

defined at /home/runner/work/CFAR/CFAR/src/lib/gaussians.jl:16.

CFAR.work_dir — Method

work_dir(args...) -> String

Summary

Points to the work directory containing raw datasets, processed datasets, and results.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

work_dir(args)

defined at /home/runner/work/CFAR/CFAR/src/lib/utils/drwatson.jl:20.